more robust tests
